fix(FEC-11431): refactor preview thumb height from real image sprite height (#473)
Following #466, #471, #472 - code was refactored so image thumb height is taken from real image sprite. This overrides all issues regarding height miscalculations and the need for videoHeight (which is missing when casting)
In addition kaltura player was passed to ThumbsManager instead of core player to fix preview image when casting as the duration when casting needs to come from cast player and not the core one.