fix(FEC-13426): Player v7| playlist| image entry cover the whole player and the user cannot navigate to another entry
Image entries in playlist should always be timed images (with the source having duration) so that the ui shows prev and next buttons for them.
In playlist manager we check for the entry type to see if it's an image, and if so, we set the source duration to imageDuration (a field of the playlist).

At the point where we test this condition, playlistItem might have an empty type field, as its value set only after it's is copied from the source after setMedia. This would fail the condition and have duration stay not set.
The fix is to check the type of the source itself instead of checking the type of the playlistItem.