feat(FEC-14415): Add option for initial HTTP GET api call to server (#944)
### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**
https://kaltura.atlassian.net/browse/FEC-14415
Players calls are being blocked by proxy, since first call is a complex
POST call,
To overcome it, we need the first call to be an HTTP GET call.

**Fix:**
User can add config - 
  provider: {
          env: {
            initCallToServer:    "/service/system/action/getversion"
          }
      }
Once this is added, the first api call to the server will be to
<serverURL>**/service/system/action/getversion/f**ormat/1
#### Resolves FEC-14415