fix(FEC-14517): Add handling for preload
### Description of the Changes

When playback.preload is active, we don't want to play the fallback
source immediately on drm error, because play() hasn't been called yet
on the original source (the original source was loaded because of the
preload, not because of play).

In case preload is not active, if the fallback source has been loaded it
means that play() was already called on the original source, so we want
to also play the fallback source.

Resolves FEC-14517